Separating lakectl test util functions and their tests
