+++ nessie/lakectl_test.go
@@ -7,22 +7,22 @@ import (
 var emptyVars = make(map[string]string)
 
 func TestLakectlHelp(t *testing.T) {
-	runCmdAndVerifySuccessWithFile(t, lakectl(), false, "lakectl_help", emptyVars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" --help", false, "lakectl_help", emptyVars)
-	runCmdAndVerifySuccessWithFile(t, lakectl(), true, "lakectl_help", emptyVars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" --help", true, "lakectl_help", emptyVars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl(), false, "lakectl_help", emptyVars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" --help", false, "lakectl_help", emptyVars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl(), true, "lakectl_help", emptyVars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" --help", true, "lakectl_help", emptyVars)
 
 }
 
 func TestBasicRepoActions(t *testing.T) {
 
-	// runCmdAndVerifySuccess(t, lakectl()+" repo list", false, "\n", emptyVars)
+	// RunCmdAndVerifySuccess(t, Lakectl()+" repo list", false, "\n", emptyVars)
 
 	// Fails due to the usage of repos for isolation - nessie creates repos in parallel and
 	// the output of 'repo list' command cannot be well defined
 	// lakectl repo list with no repo created. Verifying terminal and piped formats
-	// runCmdAndVerifySuccess(t, lakectl()+" repo list --no-color", true, "\n", emptyVars)
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list", true, "lakectl_repo_list_empty.term", emptyVars)
+	// RunCmdAndVerifySuccess(t, Lakectl()+" repo list --no-color", true, "\n", emptyVars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list", true, "lakectl_repo_list_empty.term", emptyVars)
 
 	// Create repo using lakectl repo create and verifying the output
 	// A variable mapping is used to pass random generated names for verification
@@ -33,15 +33,15 @@ func TestBasicRepoActions(t *testing.T) {
 		"STORAGE": storage,
 		"BRANCH":  mainBranch,
 	}
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" repo create lakefs://"+repoName+" "+storage, false, "lakectl_repo_create", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo create lakefs://"+repoName+" "+storage, false, "lakectl_repo_create", vars)
 
 	// lakectl repo list is expected to show the created repo
 
 	// Fails due to the usage of repos for isolation - nessie creates repos in parallel and
 	// the output of 'repo list' command cannot be well defined
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list", false, "lakectl_repo_list_1", vars)
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list --no-color", true, "lakectl_repo_list_1", vars)
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list", true, "lakectl_repo_list_1.term", vars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list", false, "lakectl_repo_list_1", vars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list --no-color", true, "lakectl_repo_list_1", vars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list", true, "lakectl_repo_list_1.term", vars)
 
 	// Create a second repo. Vars for the first repo are being saved in a new map, in order to be used
 	// for a follow-up verification with 'repo list'
@@ -52,11 +52,11 @@ func TestBasicRepoActions(t *testing.T) {
 	// }
 
 	// Trying to create the same repo again fails and does not change the list
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" repo create lakefs://"+repoName+" "+storage, false, "lakectl_repo_create_not_unique", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" repo create lakefs://"+repoName+" "+storage, false, "lakectl_repo_create_not_unique", vars)
 
 	// Fails due to the usage of repos for isolation - nessie creates repos in parallel and
 	// the output of 'repo list' command cannot be well defined
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list", false, "lakectl_repo_list_1", vars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list", false, "lakectl_repo_list_1", vars)
 
 	// Create another repo with non-default branch
 	repoName2 := generateUniqueRepositoryName()
@@ -65,7 +65,7 @@ func TestBasicRepoActions(t *testing.T) {
 	vars["REPO"] = repoName2
 	vars["STORAGE"] = storage2
 	vars["BRANCH"] = notDefaultBranchName
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" repo create lakefs://"+repoName2+" "+storage2+" -d "+notDefaultBranchName, true, "lakectl_repo_create", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo create lakefs://"+repoName2+" "+storage2+" -d "+notDefaultBranchName, true, "lakectl_repo_create", vars)
 
 	// The generated names are also added to the verification vars map
 
@@ -74,23 +74,23 @@ func TestBasicRepoActions(t *testing.T) {
 	// listVars["REPO2"] = repoName2
 	// listVars["STORAGE2"] = storage2
 	// listVars["BRANCH2"] = notDefaultBranchName
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list", false, "lakectl_repo_list_2", listVars)
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list --no-color", true, "lakectl_repo_list_2", listVars)
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list", true, "lakectl_repo_list_2.term", listVars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list", false, "lakectl_repo_list_2", listVars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list --no-color", true, "lakectl_repo_list_2", listVars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list", true, "lakectl_repo_list_2.term", listVars)
 
-	// runCmdAndVerifySuccessWithFile(t, lakectl()+" repo list --after "+repoName, false, "lakectl_repo_list_1", vars)
+	// RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo list --after "+repoName, false, "lakectl_repo_list_1", vars)
 
 	// Trying to delete a repo using malformed_uri
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" repo delete "+repoName2+" -y", false, "lakectl_repo_delete_malformed_uri", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" repo delete "+repoName2+" -y", false, "lakectl_repo_delete_malformed_uri", vars)
 
 	// Trying to delete a repo using malformed_uri, using terminal
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" repo delete "+repoName2+" -y", true, "lakectl_repo_delete_malformed_uri.term", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" repo delete "+repoName2+" -y", true, "lakectl_repo_delete_malformed_uri.term", vars)
 
 	// Deleting a repo
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" repo delete lakefs://"+repoName2+" -y", false, "lakectl_repo_delete", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo delete lakefs://"+repoName2+" -y", false, "lakectl_repo_delete", vars)
 
 	// Trying to delete again
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" repo delete lakefs://"+repoName2+" -y", false, "lakectl_repo_delete_not_found", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" repo delete lakefs://"+repoName2+" -y", false, "lakectl_repo_delete_not_found", vars)
 }
 
 func TestCommit(t *testing.T) {
@@ -101,14 +101,14 @@ func TestCommit(t *testing.T) {
 		"STORAGE": storage,
 		"BRANCH":  mainBranch,
 	}
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_404", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" repo create lakefs://"+repoName+" "+storage, false, "lakectl_repo_create", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_initial", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" fs upload -s files/ro_1k lakefs://"+repoName+"/"+mainBranch+"/ro_1k.1", false, "lakectl_fs_upload", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_initial", vars)
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" commit lakefs://"+repoName+"/"+mainBranch, false, "lakectl_commit_no_msg", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" commit lakefs://"+repoName+"/"+mainBranch+" -m \"nessie_lakectl:TestCommit\"", false, "lakectl_commit", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_with_commit", vars)
-	runCmdAndVerifyFailureWithFile(t, lakectl()+" commit lakefs://"+repoName+"/"+mainBranch+" -m \"nessie_lakectl:should fail\"", false, "lakectl_commit_no_change", vars)
-	runCmdAndVerifySuccessWithFile(t, lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_with_commit", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_404", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" repo create lakefs://"+repoName+" "+storage, false, "lakectl_repo_create", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_initial", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" fs upload -s files/ro_1k lakefs://"+repoName+"/"+mainBranch+"/ro_1k.1", false, "lakectl_fs_upload", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_initial", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" commit lakefs://"+repoName+"/"+mainBranch, false, "lakectl_commit_no_msg", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" commit lakefs://"+repoName+"/"+mainBranch+" -m \"nessie_lakectl:TestCommit\"", false, "lakectl_commit", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_with_commit", vars)
+	RunCmdAndVerifyFailureWithFile(t, Lakectl()+" commit lakefs://"+repoName+"/"+mainBranch+" -m \"nessie_lakectl:should fail\"", false, "lakectl_commit_no_change", vars)
+	RunCmdAndVerifySuccessWithFile(t, Lakectl()+" log lakefs://"+repoName+"/"+mainBranch, false, "lakectl_log_with_commit", vars)
 }
